Fix for deploy redirects
There was an error where refreshing a page would result in 404 error. This was because the app uses React Router. So when you refresh on a page that is not the root, netlify doesn't know where to send the request to. This redirect tells it to send the request to the root which has the router on it.